// apps/web/api/schemas/channelRequest.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * API Contracts for Komik
 * OpenAPI spec version: 1.0
 */
import type { ChannelRequestPayload } from './channelRequestPayload';
import type { ChannelRequestTypeOf } from './channelRequestTypeOf';

export interface ChannelRequest {
  name: string;
  typeOf: ChannelRequestTypeOf;
  payload: ChannelRequestPayload;
}